move evaluate_document_layer to evaluate_prediction.py
